moved common const, type, func into modalUtils
