+++ src/renderer/Generics/redesign/Modal/NodeSettingsModal.tsx
@@ -3,8 +3,7 @@ import { useState, useCallback } from 'react';
 import NodeSettingsWrapper from 'renderer/Presentational/NodeSettings/NodeSettingsWrapper';
 import electron from 'renderer/electronGlobal';
 import { Modal } from './Modal';
-import { ModalConfig } from './ModalManager';
-import modalOnChangeConfig from './modalOnChangeConfig';
+import { modalOnChangeConfig, ModalConfig } from './modalUtils';
 
 type Props = {
   modalOnClose: () => void;
