--- src/renderer/Generics/redesign/Modal/modalRoutes.js
@@ -1,10 +1,0 @@
-export const modalRoutes = Object.freeze({
-  addNode: 'addNode',
-  nodeSettings: 'nodeSettings',
-  preferences: 'preferences',
-  addValidator: 'addValidator',
-  clientVersions: 'clientVersions',
-  stopNode: 'stopNode',
-  removeNode: 'removeNode',
-  updateUnavailable: 'updateUnavailable',
-});
